pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "blinkcharging-be-eva-portal"
version = "0.1.0"
description = "Async Python client for the Belgian Blink Charging EVA portal (eva.blinkcharging.be)"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Michaël Arnauts" }]
license = { text = "MIT" }
keywords = ["blink", "bluecorner", "ev", "charging", "home-assistant"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Home Automation",
]
dependencies = [
    "httpx>=0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "pytest-asyncio>=0.21",
    "python-dotenv>=1.0",
    "ruff>=0.6",
]

[project.urls]
Source = "https://github.com/michaelarnauts/python-blinkcharging-be-eva-portal"

[project.scripts]
blinkcharging-dump = "blinkcharging_be_eva_portal.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/blinkcharging_be_eva_portal"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B"]

[tool.ruff.lint.isort]
known-first-party = ["blinkcharging_be_eva_portal"]
